+++ tests/dummy/app/templates/index.hbs
@@ -1,1 +1,1 @@
-{{ui-remark/render node=model.content settings=model}}
+{{ui-remark/render page=model}}
