+++ tests/dummy/app/templates/pages.hbs
@@ -1,1 +1,1 @@
-{{ui-remark/render node=model.content settings=model}}
+{{ui-remark/render page=model}}
